Allow default heights to be derived from content.
This makes it smoother to render cloaked collections where the item
heights may differ substantially from one another. Ambitious
applications may also use this ability to cache and re-use the
actual rendered height for content when moving between screens.

Adds `defaultHeightProperty` to the cloaked-collection configuration.
(See README for details.)

So that the default height can be derived from the item controller (if
any), this commit also makes it so that the item controller for each
cloaked view is created when the cloaked collection is created.
(Previously, the item controller was not created until the cloaked view
was uncloaked.)